modules/packetio: reorganize dpdk callback files
Move all dpdk port feature files into a dedicated directory. They're
already in their own namespace and this layout mirrors that.

No functional change.
