+++ src/modules/packetio/drivers/dpdk/port/mac_filter.cpp
@@ -3,7 +3,7 @@
 
 #include "packet/type/mac_address.hpp"
 #include "packetio/drivers/dpdk/model/port_info.hpp"
-#include "packetio/drivers/dpdk/port_filter.hpp"
+#include "packetio/drivers/dpdk/port/filter.hpp"
 
 namespace openperf::packetio::dpdk::port {
