+++ src/modules/packetio/drivers/dpdk/port/signature_encoder.cpp
@@ -4,8 +4,8 @@
 #include "packetio/drivers/dpdk/dpdk.h"
 #include "packetio/drivers/dpdk/model/port_info.hpp"
 #include "packetio/drivers/dpdk/mbuf_signature.hpp"
-#include "packetio/drivers/dpdk/port_signature_encoder.hpp"
-#include "packetio/drivers/dpdk/port_signature_utils.hpp"
+#include "packetio/drivers/dpdk/port/signature_encoder.hpp"
+#include "packetio/drivers/dpdk/port/signature_utils.hpp"
 #include "spirent_pga/api.h"
 #include "units/data-rates.hpp"
 #include "utils/soa_container.hpp"
