+++ src/modules/packetio/drivers/dpdk/port/timestamper.cpp
@@ -3,7 +3,7 @@
 #include "core/op_log.h"
 #include "packetio/drivers/dpdk/dpdk.h"
 #include "packetio/drivers/dpdk/model/port_info.hpp"
-#include "packetio/drivers/dpdk/port_timestamper.hpp"
+#include "packetio/drivers/dpdk/port/timestamper.hpp"
 #include "timesync/chrono.hpp"
 #include "units/data-rates.hpp"
 
@@ -91,8 +91,8 @@ callback_timestamper::callback_timestamper(
     , m_port(other.m_port)
 {}
 
-callback_timestamper& callback_timestamper::
-operator=(callback_timestamper&& other) noexcept
+callback_timestamper&
+callback_timestamper::operator=(callback_timestamper&& other) noexcept
 {
     if (this != &other) {
         m_callbacks = std::move(other.m_callbacks);
