+++ src/modules/packetio/workers/dpdk/worker_controller.hpp
@@ -10,12 +10,12 @@
 #include "packetio/generic_driver.hpp"
 #include "packetio/generic_event_loop.hpp"
 #include "packetio/generic_workers.hpp"
-#include "packetio/drivers/dpdk/port_filter.hpp"
-#include "packetio/drivers/dpdk/port_packet_type_decoder.hpp"
-#include "packetio/drivers/dpdk/port_signature_decoder.hpp"
-#include "packetio/drivers/dpdk/port_signature_encoder.hpp"
-#include "packetio/drivers/dpdk/port_timestamper.hpp"
 #include "packetio/workers/dpdk/callback.hpp"
+#include "packetio/drivers/dpdk/port/filter.hpp"
+#include "packetio/drivers/dpdk/port/packet_type_decoder.hpp"
+#include "packetio/drivers/dpdk/port/signature_decoder.hpp"
+#include "packetio/drivers/dpdk/port/signature_encoder.hpp"
+#include "packetio/drivers/dpdk/port/timestamper.hpp"
 #include "packetio/workers/dpdk/port_feature_controller.hpp"
 #include "packetio/workers/dpdk/tx_scheduler.hpp"
 #include "packetio/workers/dpdk/worker_api.hpp"
